Updated unused loading screen
