+++ app/(dashboard)/(routes)/page.tsx
@@ -1,0 +1,41 @@
+"use client";
+
+import { ArrowRight } from "lucide-react";
+import { useRouter } from "next/navigation";
+import { Card } from "@/components/ui/card";
+import { cn } from "@/lib/utils";
+import { tools } from "@/constants";
+
+export default function HomePage() {
+  const router = useRouter();
+
+  return (
+    <div className="min-h-0 flex-1 overflow-y-auto">
+      <div className="my-8 space-y-4">
+        <h2 className="text-center text-2xl font-bold md:text-4xl">
+          Welcome to Automated Mp3 Tag Editor
+        </h2>
+        <p className="text-center text-sm font-light text-muted-foreground md:text-lg">
+          Explore the power of AI-Driven Music Metadata Aggregation
+        </p>
+      </div>
+      <div className="space-y-4 px-4 pb-8 md:px-20 lg:px-32">
+        {tools.map((tool) => (
+          <Card
+            onClick={() => router.push(tool.href)}
+            key={tool.href}
+            className="flex cursor-pointer items-center justify-between border-black/5 p-4 transition hover:shadow-md"
+          >
+            <div className="flex items-center gap-x-4">
+              <div className={cn("w-fit rounded-md p-2", tool.bgColor)}>
+                <tool.icon className={cn("h-8 w-8", tool.color)} />
+              </div>
+              <div className="font-semibold">{tool.label}</div>
+            </div>
+            <ArrowRight className="h-5 w-5" />
+          </Card>
+        ))}
+      </div>
+    </div>
+  );
+}
