+++ next.config.js
@@ -5,15 +5,3 @@ const nextConfig = {
 }
 
 module.exports = nextConfig
-
-// module.exports = {
-//   async redirects() {
-//     return [
-//       {
-//         source: '/',
-//         destination: '/dashboard',
-//         permanent: true,
-//       },
-//     ]
-//   },
-// }
